refactoring, templating
