+++ templates/404.html
@@ -1,0 +1,9 @@
+{% extends "layout.html" %}
+{% block title %}
+404, page not found
+{% endblock %}
+{% block content %}
+<h1>404</h1>
+<p>The page you're looking for does not exist.</p>
+<p><a href="{{ url_for('index') }}">Return home</a></p>
+{% endblock %}
