+++ templates/index.html
@@ -2,6 +2,9 @@
 {% block title %}
 Configurator
 {% endblock %}
+{% block stylesheet %}
+{{ url_for('static', filename='index.css') }}
+{% endblock %}
 {% block content %}
 <div class="simple-mode">
     <form action="update/simple" method="POST">
@@ -34,7 +37,5 @@ Configurator
         <input type="submit" value="Update">
     </form>
 </div>
-<div>
-    <a href="{{ url_for('advanced') }}">Advanced</a>
-</div>
+<p>Last updated {{ now.strftime("%d %b %Y at %X") }}. Reboot for updated values.</p>
 {% endblock %}
